+++ config/export.yaml
@@ -4,6 +4,13 @@ services:
     autoconfigure: true
     public: false
 
+  # controllers are imported separately to make sure they're public
+  # and have a tag that allows actions to type-hint services
+  Pimcore\Bundle\StudioBackendBundle\Export\Controller\:
+    resource: '../src/Export/Controller'
+    public: true
+    tags: [ 'controller.service_arguments' ]
+
   #
   # Handler
   #
@@ -16,6 +23,8 @@ services:
 
   Pimcore\Bundle\StudioBackendBundle\Export\ExportServiceInterface: ~
   Pimcore\Bundle\StudioBackendBundle\Export\Csv\CsvExportService: ~
+  Pimcore\Bundle\StudioBackendBundle\Export\Service\DownloadServiceInterface:
+    class: Pimcore\Bundle\StudioBackendBundle\Export\Service\DownloadService
 
   #
   # Event Subscriber
